doing whatever I want bitch

+++ StockTrader/src/main/java/com/restResource/StockTrader/entity/logging/UserCommandLog.java
@@ -31,5 +31,5 @@ public class UserCommandLog {
     @XmlElement
     String filename; //can be null
     @XmlElement
-    Integer funds; //can be null
+    Double funds; //can be null
 }
